+++ test/test_helper.exs
@@ -18,16 +18,26 @@ if :gmail in include do
 end
 
 if :mailpit in include do
-  case :httpc.request(:get, {~c"http://localhost:8025/api/v1/info", []}, [], []) do
-    {:ok, {{_version, _status = 200, _reason}, _headers, _body}} ->
-      :ok
+  mailpit_api_ports = [8025, 8026, 8027]
 
-    {:error, {:failed_connect, [_to_address, {:inet, [:inet], :econnrefused}]}} ->
-      shell.error("""
-      To enable Mailpit tests, start the local container with the following command:
+  unavailable_ports =
+    Enum.reject(mailpit_api_ports, fn port ->
+      url = String.to_charlist("http://localhost:#{port}/api/v1/info")
 
-          docker run -d --rm -p 1025:1025 -p 8025:8025 -e "MP_SMTP_AUTH_ACCEPT_ANY=1" -e "MP_SMTP_AUTH_ALLOW_INSECURE=1" --name mailpit axllent/mailpit
-      """)
+      match?(
+        {:ok, {{_version, _status = 200, _reason}, _headers, _body}},
+        :httpc.request(:get, {url, []}, [], [])
+      )
+    end)
+
+  if unavailable_ports != [] do
+    shell.error("""
+    To enable Mailpit tests, start the local services:
+
+        docker compose up -d --wait
+
+    Unavailable Mailpit API ports: #{Enum.join(unavailable_ports, ", ")}
+    """)
   end
 end